+++ CHANGELOG.md
@@ -35,6 +35,9 @@ to use the newest tag with new release
 - Added availability to import steps by scenario name. Added some default scenarios.
   Added availability to create custom scenarios.
 - availability to use `tasks_from` to import any step
+- `zone` variable to edit instance zone
+- `edit_topology_timeout` variable to wait until cluster become healthy
+  after editing topology
 
 ### Changed